i fixed float error

+++ stats_functions.py
@@ -6,8 +6,8 @@ def med(a_list):
     if len(sorted_list)%2 ==1:
         middle = len(sorted_list)/2
         return sorted_list[middle]
-    m1 = (len(sorted_list)-1)/2
-    m2 = (len(sorted_list)+1)/2
+    m1 = int((len(sorted_list)-1)/2)
+    m2 = int((len(sorted_list)+1)/2)
     return (sorted_list[m1] + sorted_list[m2])/2
 
 def range(a_list):
